Add image generation script and 2 master images
The diagonal master image is slightly improved compared
to the existing images.
A 1-pixel border is also added around the pictures, to make them
display better in the client
The color of steel wrenches is also lightened by the script,
so that it is more similar to the other steel tools.

The actual pictures of the steel wrench were not updated yet.
(See the next commit)