Added a player-relative positioning mode.
The existing (absolute) positioning mode sets a fixed orientation
relative to the world.
The new relative mode, sets the orientation relative to the player
(or specifically, relative to the punched side of the node).

E.g.:
In absolute mode, if a wrench is configured using a chest with
the lock-side facing north, and the lid facing east. then every subsequent
chest that is positioned, will get its lock-side facing north, and its
lid east.

In relative mode, if a wrench is configured using a chest, by punching
the side with the lock, while the lid is to the right, then every
subsequent chest that is positioned, will get its lock-side on the
side the player punched, and the lid to the right of the player.